backend_fastapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
ccxt
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
websockets
redis
pandas
alembic
